+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/tarahussain/Desktop/github-react-app/src/index.js":"1","/Users/tarahussain/Desktop/github-react-app/src/App.js":"2","/Users/tarahussain/Desktop/github-react-app/src/components/SearchBar.js":"3","/Users/tarahussain/Desktop/github-react-app/src/components/Results.js":"4"},{"size":219,"mtime":1606839696558,"results":"5","hashOfConfig":"6"},{"size":267,"mtime":1606839696547,"results":"7","hashOfConfig":"6"},{"size":873,"mtime":1606840479485,"results":"8","hashOfConfig":"6"},{"size":698,"mtime":1606840525591,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"s7pyka",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/tarahussain/Desktop/github-react-app/src/index.js",[],["19","20"],"/Users/tarahussain/Desktop/github-react-app/src/App.js",[],"/Users/tarahussain/Desktop/github-react-app/src/components/SearchBar.js",[],"/Users/tarahussain/Desktop/github-react-app/src/components/Results.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/tarahussain/Desktop/github-react-app/src/index.js":"1","/Users/tarahussain/Desktop/github-react-app/src/App.js":"2","/Users/tarahussain/Desktop/github-react-app/src/components/index.js":"3","/Users/tarahussain/Desktop/github-react-app/src/components/results/index.js":"4","/Users/tarahussain/Desktop/github-react-app/src/components/searchbar/index.js":"5"},{"size":219,"mtime":1606839696558,"results":"6","hashOfConfig":"7"},{"size":261,"mtime":1606843137445,"results":"8","hashOfConfig":"7"},{"size":112,"mtime":1606843205107,"results":"9","hashOfConfig":"7"},{"size":1972,"mtime":1606845255513,"results":"10","hashOfConfig":"7"},{"size":874,"mtime":1606843220666,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"s7pyka",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/tarahussain/Desktop/github-react-app/src/index.js",[],["23","24"],"/Users/tarahussain/Desktop/github-react-app/src/App.js",[],"/Users/tarahussain/Desktop/github-react-app/src/components/index.js",[],"/Users/tarahussain/Desktop/github-react-app/src/components/results/index.js",[],"/Users/tarahussain/Desktop/github-react-app/src/components/searchbar/index.js",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
