+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as SearchBar } from "./searchbar/index";
+export { default as Results } from "./results/index";
