+++ src/components/results/results.css
@@ -1,0 +1,41 @@
+* {
+  box-sizing: border-box;
+}
+
+.repoContainer {
+  border: 3px solid black;
+  border-radius: 10px;
+  margin: 10px;
+  /* padding: 10px;
+  display: flex;
+  flex-direction: column;
+  align-items: center; */
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2); /* this adds the "card" effect */
+  padding: 16px;
+  text-align: center;
+  background-color: #f1f1f1;
+}
+
+#column {
+  float: left;
+  width: 31%;
+  padding: 0 10px;
+}
+
+.row {
+  margin: 0 -5px;
+}
+
+.row:after {
+  content: "";
+  display: table;
+  clear: both;
+}
+
+@media screen and (max-width: 600px) {
+  #column {
+    width: 100%;
+    display: block;
+    margin-bottom: 20px;
+  }
+}
